api_imagenes/requirements.txt
fastapi
uvicorn
aio_pika
pillow
python-multipart

